Module metafactory::aggregate

Implements a factory that aggregates the results of other factories of the same type.

Structs

Aggregate
Proxy for initializing aggregate factory without caring about the type used.