Trait metafactory::ToMetaFactoryUnstable

pub trait ToMetaFactory {
    fn to_metafactory<'a>(self) -> Box<MetaFactory + 'a>;
}

Trait for values convertable to MetaFactory.

This trait is implemented for values that can be used as sources for object creation.

Required Methods

fn to_metafactory<'a>(self) -> Box<MetaFactory + 'a>

Creates a MetaFactory that has information about object constructor: produced object type, argument types, and a method to get this getter.